backend/requirements.txt
Django==4.2.3
django-colorfield==0.9.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.7.0
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
requests==2.26.0
reportlab==4.0.4
psycopg2